@@ -17,10 +17,11 @@
     "web/"
   ],
   "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^5.4.0",
     "babel": "^6.23.0",
     "babel-eslint": "^10.0.1",
     "body-parser": "^1.19.0",
-    "eslint": "^5.16.0",
+    "eslint": "7.1.0",
     "eslint-config-prettier": "^4.2.0",
     "eslint-plugin-import": "^2.17.2",
     "eslint-plugin-node": "^9.0.1",
