@@ -1,4 +1,7 @@
-module.exports.ping = async event => {
+import log from '@shared/logger'
+
+export const ping = async event => {
+  log('info', event)
   return {
     statusCode: 200,
     body: JSON.stringify(
